Modernize Battleship desktop app
