+++ battle/src/battleship/PlayerBoard.java
@@ -1,5 +1,10 @@
-import java.awt.*;
-import java.awt.event.*;
+package battleship;
+
+import java.awt.Canvas;
+import java.awt.Color;
+import java.awt.Graphics;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
 
 public class PlayerBoard extends Canvas implements MouseListener
 {
